backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Supabase
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-docx>=0.8.11
nltk>=3.8.0
numpy>=1.24.0

# Gmail API integration
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
python-dateutil>=2.8.0

# Job scraping
python-jobspy>=1.1.4